Refactored skill and magic effect creation
* Added an enumeration for skill numbers
* Extracted magic effect initialization stuff into separate classes
* Found and fixed an issue with magic effects which are just internal - they are not sent to the client anymore. I needed to change the datatype of MagicEffectDefinition.Number from byte to short in this case. Negative numbers means that an effect is internal.

+++ src/DataModel/Configuration/MagicEffectDefinition.cs
@@ -16,8 +16,9 @@ namespace MUnique.OpenMU.DataModel.Configuration
         /// </summary>
         /// <remarks>
         /// This number is a reference for the game client.
+        /// Negative numbers are for internal usage and their effects are not meant to be exposed to the game client.
         /// </remarks>
-        public byte Number { get; set; }
+        public short Number { get; set; }
 
         /// <summary>
         /// Gets or sets the name.
